+++ app/routes/signup.tsx
@@ -4,6 +4,7 @@ import IconPassword from "~/components/icons/iconPassword";
 import IconProfile from "~/components/icons/iconProfile";
 import { photoPath } from "~/server/config.server";
 import { UserAPI } from "~/server/repository";
+import prefetchImage from "~/server/services/imagePrefetcher";
 
 export async function action({ request }: ActionFunctionArgs) {
   const formData = await request.formData();
@@ -40,12 +41,13 @@ export async function action({ request }: ActionFunctionArgs) {
 
 export async function loader({ request }: LoaderFunctionArgs) {
   const Photo = photoPath();
-  return { Photo };
+  const dogPhoto = await prefetchImage(Photo + "dog-in-the-air.jpg");
+  return { Photo, dogPhoto };
 }
 
 export default function SignUpView() {
   const actionData = useActionData<typeof action>() || null;
-  const { Photo } = useLoaderData<typeof loader>();
+  const { Photo, dogPhoto } = useLoaderData<typeof loader>();
 
   return (
     <div
@@ -200,7 +202,7 @@ export default function SignUpView() {
           </div>
         </Form>
         {/* Image */}
-        <img src={Photo + "dog-in-the-air.jpg"} />
+        <img src={dogPhoto} />
       </div>
     </div>
   );
