+++ app/server/repositories/productRepository.ts
@@ -7,23 +7,33 @@ const apiPath = `${Domain}/product`;
 export default class ProductAPI {
 
 	static async getAll(): Promise<product[]> {
-		const res = await fetch(`${apiPath}/getAll`, { method: "GET" });
-		const data = await res.json();
+		try {
+			const res = await fetch(`${apiPath}/getAll`, { method: "GET" });
+			const data = await res.json();
 
 		if (!res.ok) {
-			throw new Error(`Failed to fetch product category: ${res.status} ${res.statusText}`);
+				throw new Error(`Failed to fetch product category: ${res.status} ${res.statusText}`);
+			}
+			return data;
+		} catch (error) {
+			console.error('Error fetching products:', error);
+			throw error;
 		}
-		return data;
 	}
 
 	static async getByID(id: number): Promise<product> {
-		const res = await fetch(`${apiPath}/getById/${id}`, { method: "GET" });
-		const data = await res.json();
+		try {
+			const res = await fetch(`${apiPath}/getById/${id}`, { method: "GET" });
+			const data = await res.json();
 
 		if (!res.ok) {
-			throw new Error(`Failed to fetch product category: ${res.status} ${res.statusText}`);
+				throw new Error(`Failed to fetch product category: ${res.status} ${res.statusText}`);
+			}
+			return data;
+		} catch (error) {
+			console.error('Error fetching product by ID:', error);
+			throw error;
 		}
-		return data;
 	}
 
 	static async createProduct(
@@ -34,9 +44,10 @@ export default class ProductAPI {
 		price: number,
 		stock: number,
 	) {
-		const res = await fetch(`${apiPath}/createProduct`, {
-			method: "POST",
-			body: JSON.stringify({
+		try {
+			const res = await fetch(`${apiPath}/createProduct`, {
+				method: "POST",
+				body: JSON.stringify({
 				product_category_id: product_category_id,
 				name: name,
 				imageurl: imageurl,
@@ -48,9 +59,13 @@ export default class ProductAPI {
 
 		const data = await res.json();
 		if (!res.ok) {
-			throw new Error(`Failed to fetch product category: ${res.status} ${res.statusText}`);
+				throw new Error(`Failed to fetch product category: ${res.status} ${res.statusText}`);
+			}
+			return data;
+		} catch (error) {
+			console.error('Error creating product:', error);
+			throw error;
 		}
-		return data;
 	}
 
 	static async updateProduct(
@@ -62,9 +77,10 @@ export default class ProductAPI {
 		price: number,
 		stock: number,
 	) {
-		const res = await fetch(`${apiPath}/updateProduct`, {
-			method: "POST",
-			body: JSON.stringify({
+		try {
+			const res = await fetch(`${apiPath}/updateProduct`, {
+				method: "POST",
+				body: JSON.stringify({
 				id: id,
 				product_category_id: product_category_id,
 				name: name,
@@ -80,10 +96,15 @@ export default class ProductAPI {
 			throw new Error(`Failed to fetch product category: ${res.status} ${res.statusText}`);
 		}
 		return data;
+		} catch (error) {
+			console.error('Error updating product:', error);
+			throw error;
+		}
 	}
 
 	static async deleteProduct(id: number) {
-		const res = await fetch(`${apiPath}/deleteProduct`, {
+		try {
+			const res = await fetch(`${apiPath}/deleteProduct`, {
 			method: "DELETE",
 			body: JSON.stringify({
 				id: id,
@@ -92,8 +113,12 @@ export default class ProductAPI {
 
 		const data = await res.json();
 		if (!res.ok) {
-			throw new Error(`Failed to fetch product category: ${res.status} ${res.statusText}`);
+				throw new Error(`Failed to fetch product category: ${res.status} ${res.statusText}`);
+			}
+			return data;
+		} catch (error) {
+			console.error('Error deleting product:', error);
+			throw error;
 		}
-		return data;
 	}
 }
